refactor: remove API from component to separate file to provide bettet sestability for tests
